fix date time issue in sprint 2

+++ src/main/java/com/asg/finance/entity/ApPurchaseInvoiceHdrEntity.java
@@ -150,7 +150,7 @@ public class ApPurchaseInvoiceHdrEntity extends BaseEntity {
     private BigDecimal supplierInvAmount;
 
     @Column(name = "ROUNDING_AMOUNT")
-    private Long roundingAmount;
+    private BigDecimal roundingAmount;
 
     @Column(name = "BILL_TYPE")
     private String billType;
